Add organizer view to display full draw mapping
Add TelaVisualizarSorteioGeral to show all pairs for an event. Implement listar_por_evento in SorteioRepository and
ver_mapeamento_geral in SorteioController. Prevent performing a draw if event status is not PREPARANDO. Update App and
TelaListaSorteios navigation to use "acao_sorteio" and route to either the draw screen or the mapping view accordingly.